# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/usuarios":{"get":{"tags":["usuario-controller"],"operationId":"getAllUsuarios","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUsuario"}}}}}},"post":{"tags":["usuario-controller"],"operationId":"createUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usuario"}}}}}}},"/api/recomendacoes":{"get":{"tags":["recomendacao-controller"],"operationId":"getAllRecomendacoes","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRecomendacao"}}}}}},"post":{"tags":["recomendacao-controller"],"operationId":"createRecomendacao","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recomendacao"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recomendacao"}}}}}}},"/api/dispositivos":{"get":{"tags":["dispositivo-controller"],"operationId":"getAllDispositivos","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageDispositivo"}}}}}},"post":{"tags":["dispositivo-controller"],"operationId":"createDispositivo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Dispositivo"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dispositivo"}}}}}}},"/api/consumos":{"get":{"tags":["consumo-controller"],"operationId":"getAllConsumos","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageConsumo"}}}}}},"post":{"tags":["consumo-controller"],"operationId":"createConsumo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Consumo"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Consumo"}}}}}}},"/api/configuracoes":{"get":{"tags":["configuracao-controller"],"operationId":"getAllConfiguracoes","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageConfiguracao"}}}}}},"post":{"tags":["configuracao-controller"],"operationId":"createConfiguracao","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Configuracao"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Configuracao"}}}}}}},"/api/usuarios/{id}":{"get":{"tags":["usuario-controller"],"operationId":"getUsuarioById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usuario"}}}}}},"delete":{"tags":["usuario-controller"],"operationId":"deleteUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/recomendacoes/{id}":{"get":{"tags":["recomendacao-controller"],"operationId":"getRecomendacaoById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recomendacao"}}}}}},"delete":{"tags":["recomendacao-controller"],"operationId":"deleteRecomendacao","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/dispositivos/{id}":{"get":{"tags":["dispositivo-controller"],"operationId":"getDispositivoById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dispositivo"}}}}}},"delete":{"tags":["dispositivo-controller"],"operationId":"deleteDispositivo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/consumos/{id}":{"get":{"tags":["consumo-controller"],"operationId":"getConsumoById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Consumo"}}}}}},"delete":{"tags":["consumo-controller"],"operationId":"deleteConsumo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/configuracoes/{id}":{"get":{"tags":["configuracao-controller"],"operationId":"getConfiguracaoById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Configuracao"}}}}}},"delete":{"tags":["configuracao-controller"],"operationId":"deleteConfiguracao","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Usuario":{"required":["email","nome","senha","tipoUsuario"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"email":{"type":"string"},"senha":{"type":"string"},"dataCadastro":{"type":"string","format":"date-time"},"tipoUsuario":{"type":"string"}}},"Recomendacao":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descricao":{"type":"string"},"dataCriacao":{"type":"string","format":"date"},"tipoRecomendacao":{"type":"string"},"usuario":{"$ref":"#/components/schemas/Usuario"}}},"Dispositivo":{"required":["nome","status","tipo"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"tipo":{"type":"string"},"potencia":{"type":"number","format":"double"},"status":{"type":"string"},"usuario":{"$ref":"#/components/schemas/Usuario"}}},"Consumo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dataHora":{"type":"string","format":"date-time"},"valorConsumo":{"type":"number","format":"double"},"dispositivo":{"$ref":"#/components/schemas/Dispositivo"},"usuario":{"$ref":"#/components/schemas/Usuario"}}},"Configuracao":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tarifaEnergia":{"type":"number","format":"double"},"tempoConsumo":{"type":"integer","format":"int32"},"usuario":{"$ref":"#/components/schemas/Usuario"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageUsuario":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Usuario"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"PageRecomendacao":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Recomendacao"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageDispositivo":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Dispositivo"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageConsumo":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Consumo"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageConfiguracao":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Configuracao"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}}}}